Show switch stream options for shared files
