+++ mobile/lib/ui/viewer/file/video_stream_change.dart
@@ -1,0 +1,119 @@
+import "dart:async";
+
+import "package:flutter/material.dart";
+import "package:photos/core/event_bus.dart";
+import "package:photos/events/preview_updated_event.dart";
+import "package:photos/generated/l10n.dart";
+import "package:photos/models/file/file.dart";
+import "package:photos/models/preview/preview_item.dart";
+import "package:photos/services/filedata/filedata_service.dart";
+import "package:photos/services/preview_video_store.dart";
+import "package:photos/theme/colors.dart";
+
+class VideoStreamChangeWidget extends StatefulWidget {
+  const VideoStreamChangeWidget({
+    super.key,
+    required bool showControls,
+    required this.file,
+    required this.onStreamChange,
+    this.isPreviewPlayer = false,
+  }) : _showControls = showControls;
+
+  final bool _showControls;
+  final EnteFile file;
+  final bool isPreviewPlayer;
+  final void Function()? onStreamChange;
+
+  @override
+  State<VideoStreamChangeWidget> createState() =>
+      _VideoStreamChangeWidgetState();
+}
+
+class _VideoStreamChangeWidgetState extends State<VideoStreamChangeWidget> {
+  StreamSubscription? previewSubscription;
+  late PreviewItem? preview =
+      PreviewVideoStore.instance.previews[widget.file.uploadedFileID];
+
+  @override
+  void initState() {
+    super.initState();
+    previewSubscription =
+        Bus.instance.on<PreviewUpdatedEvent>().listen((event) {
+      final newPreview = event.items[widget.file.uploadedFileID];
+      if (newPreview != preview) {
+        setState(() {
+          preview = event.items[widget.file.uploadedFileID];
+        });
+      }
+    });
+  }
+
+  @override
+  void dispose() {
+    previewSubscription?.cancel();
+    super.dispose();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    final bool isPreviewAvailable = widget.file.uploadedFileID != null &&
+        (FileDataService.instance.previewIds
+                ?.containsKey(widget.file.uploadedFileID) ??
+            false);
+    if (!isPreviewAvailable) {
+      return const SizedBox();
+    }
+    return Align(
+      alignment: Alignment.centerRight,
+      child: AnimatedOpacity(
+        duration: const Duration(
+          milliseconds: 200,
+        ),
+        curve: Curves.easeInQuad,
+        opacity: widget._showControls ? 1 : 0,
+        child: Padding(
+          padding: const EdgeInsets.only(
+            right: 10,
+            bottom: 4,
+          ),
+          child: GestureDetector(
+            onTap: widget.onStreamChange,
+            child: Container(
+              padding: const EdgeInsets.symmetric(
+                horizontal: 8,
+                vertical: 4,
+              ),
+              decoration: BoxDecoration(
+                color: Colors.green,
+                borderRadius: const BorderRadius.all(
+                  Radius.circular(200),
+                ),
+                border: Border.all(
+                  color: strokeFaintDark,
+                  width: 1,
+                ),
+              ),
+              child: Row(
+                mainAxisAlignment: MainAxisAlignment.end,
+                mainAxisSize: MainAxisSize.min,
+                children: [
+                  const Icon(Icons.play_arrow, size: 16),
+                  const SizedBox(width: 2),
+                  Text(
+                    widget.isPreviewPlayer
+                        ? S.of(context).playOriginal
+                        : S.of(context).playStream,
+                    style: const TextStyle(
+                      fontSize: 12,
+                      fontWeight: FontWeight.w600,
+                    ),
+                  ),
+                ],
+              ),
+            ),
+          ),
+        ),
+      ),
+    );
+  }
+}
