+++ mobile/lib/ui/viewer/file/video_widget_native.dart
@@ -28,8 +28,8 @@ import "package:photos/ui/common/loading_widget.dart";
 import "package:photos/ui/notification/toast.dart";
 import "package:photos/ui/viewer/file/native_video_player_controls/play_pause_button.dart";
 import "package:photos/ui/viewer/file/native_video_player_controls/seek_bar.dart";
-import "package:photos/ui/viewer/file/preview_status_widget.dart";
 import "package:photos/ui/viewer/file/thumbnail_widget.dart";
+import "package:photos/ui/viewer/file/video_stream_change.dart";
 import "package:photos/utils/dialog_util.dart";
 import "package:photos/utils/exif_util.dart";
 import "package:photos/utils/file_util.dart";
@@ -357,7 +357,7 @@ class _VideoWidgetNativeState extends State<VideoWidgetNative>
                                 ValueListenableBuilder(
                                   valueListenable: _showControls,
                                   builder: (context, value, _) {
-                                    return PreviewStatusWidget(
+                                    return VideoStreamChangeWidget(
                                       showControls: value,
                                       file: widget.file,
                                       isPreviewPlayer: widget.selectedPreview,
